Renaming utils to prevent naming collisions
