+++ dota_utils/functions.py
