+++ setup.cfg
@@ -1,4 +1,5 @@
 [flake8]
+# Cyclomatic complexity
 max-complexity = 10
 max-line-length = 120
 exclude =
